Added tests to ensure the combination of items works correctly. (This becomes relevant once guard clauses etc. are introduced)
Honest note: In a normal working environment I would consult a teammate whether this is something to be tested here or possibly somewhere else. I'd rather ask a question too much than leave something uncovered or tested badly. For this exercise I decided to add the tests as they become relevant for the commit to follow this one.