Added Smelly Items
Note: ItemUpdateQualityStrategyFactory has a switch statement with hardcoded Strings. For now I decided to keep it like this, because of the requirement of not touching the Item class. Normally I would like to have a type in Item which can be used to pattern match in this switch statement which would be cleaner.

I personally still prefer the current switch statement over "if, else-if, else-if, else-if, else"